requirements.txt
Flask>=3.0,<4.0
PyJWT>=2.9,<3.0
Werkzeug>=3.0,<4.0
mlx-lm>=0.31,<0.32
